Implement iteration of StructureBSP block
And with that I believe I have whatever I care about so far in the
scenario tag that allows me to extract the BSP and some key things that
I actually care about. Interpreting the BSP itself is going to be a lot
of work but I'm now capable of interpreting everything up to loading the
map file and interpreting the base-tag.