Fix low-resolution mipmap streaming
When the size of the current image itself is smaller than the resolution
of the block-size. Then it still takes up just one block. Previously all
of the lower mipmaps(2x2, 2x1, etc) were resolving to a size of "0"
causing them to be left as garbage and skipped in the stream buffer
upload. Fixes visual artifacts for when the camera is very far.