Add `DescriptorUpdateBatch` utility structure
Allows batch-processing of descriptor-writes and descriptor-copies to
avoid having an unbounded amount of space when generating a bunch of
writes. Now it will flush regularly and utilize a finite amount of
re-usable space rather than allocate thousands upon thousands of
descriptor writes and write-info structures.

It is likely that whatever shader-manager structure I end up making will
be utilizing this.