Implement reusable sampler cache
Introduces a new `SamplerCache` object for reusable sampler objects.
Also introduces a Scene descriptor set at descriptor set slot 0 with
globally-accessable sampler objects. This allows for an easy abstraction
for how the blam engine itself interfaces with its DirectX9 sampler
objects. Some shader-tags use an unfiltered sampler tag. Rather than
having a redundant amount of descriptors pointing to the variations.
Each of the samplers can be made globally available and picked from at
runtime. Might be slower to branch in such a way but this is a good
pattern to introduce either way while a formal shader system is made.