Implement a `TagVisitor`
Implements a better pattern for visiting tags that allows for some sort
of task-graph to be used so that multiple visitors can possibly happen
in order and within a certain dependency of each other. The current
use-case in mind is having a bitmap loader run first, then having a
shader-loader run after the bitmap loader has run so that it can
immediately create descriptor sets that reference those bitmaps. These
systems will declare their dependency and the `DispatchTagVisitors`
function will resolve these dependencies with a simple array-based
directed acyclic graph.