Partition device list by perferred criteria
Rather than just picking the first discrete GPU, we partition the list
of available devices and then pick the "best" out of the partitioned
list. So even if none of them pass our preferred criteria, or some of
them pass more critiera than others, we always end up with some kind of
device to use.