Migrate bitmap Blam-VK conversons to Format.hpp
Centralize all of these bitmap-trait decisions into an internal
`FormatTraits.hpp` file. The weirdest format I've had to handle so far
is A4R4G4B4, which is mapped to R4G4B4A4 and utilizes a component
remapping. Stuff like this will probably happen later on with some of
the other formats like Y and AY and P8.