Display BSP-Material Surface Span
BSP materials will seem to claim regions of the surface-heap. At
run-time it will test all of the sub-clusters found within each cluster
against an AABB as a form of occlusion testing and will set a bit-array
for each surface that passes the occlusion test. The bit-array is an
array of 32-bit integers where each surface-index is assigned a bit
within each 32-bit integer to designate that subcluster as being "on".