Add `GetBlock` for TagBlock iteration
Rather than pass in a pointer and virtual base on each `GetSpan`
function, the immediate structures with a virtual base-address will
instead be the ones to resolve TagBlocks into actual data. The next and
ideal iteration of this is to implement a sort of `VirtualHeap` type
that encapsulates a virtual-base-address and a pointer to file-memory to
effectively "de-virtualize" these virtual-addresses into file
addresses.