Refactor framebuffer download buffer
This used to be a general-purpose staging buffer, but now it is just a
"download buffer" that gets a render-texture written into it. It now
exhaust a list of candidate download-buffer memory-types and ensures
that the memory is ready to be read by the host with a call to
`vkInvalidateMappedMemoryRanges`.