Draft `Swapchain` vulkan object
Hoping to be able to abstract away the details of creating and recreating a swapchain with settings relating to vsync and eventually allowing further configurations for HDR and simplifying user configurations related to latency.
Generally, SDL should not leak into any of these implementation details so allow these classes to be reused for other projects too.