Draft swapchain rendering
First pass of utilizing `vk_ext_headless_surface` to verify that swapchain creation and usage is sound for the current single-frame rendering use-case.

Generally SDL stuff should be isolated in such a way that all we want out of it, as far as vulkan is concerned, is a surface-object and nothing else.